Add usb support and repro exp

+++ scripts/bench_usb.sh
@@ -5,16 +5,23 @@ targets=(
 	'mwifiex_usb'
 	'rsi_usb'
 )
+
+for i in `seq 7`; do
 for t in ${targets[@]}; do
-	echo $t
+	if [ -d /data/$USER/drifuzz-usb/$t-$i ]; then
+		continue
+	fi
+	date
+	echo $t $i
 	screen -S fuzz -d -m scripts/run_conc_model.sh --usb $t
 	sleep 1h
 	scripts/kill.sh
 	pkill screen
+	echo -n "coverage: "
+	tail -n 1 work/work-$t-conc-model/evaluation/data.csv |awk -F';' '{print $17}'
+	mkdir -p /data/$USER/drifuzz-usb/$t-$i
+	mv work/work-$t-conc-model/ /data/$USER/drifuzz-usb/$t-$i
+done
 done
 
-for t in ${targets[@]}; do
-	echo $t
-	tail -n 1 work/work-$t-conc-model/evaluation/data.csv
-done
